move availble templates to the actions
